Adapting to VTK double bounds instead of float. Note: in vtkDesktopDeliveryClient I have a really lame conversion from double to float before calling vtkCubeSource->SetBounds(). Seems like vtkCubeSource->SetBounds should be changed to take doubles.

+++ ParaView/ParaView/vtkDesktopDeliveryClient.h
@@ -65,7 +65,7 @@ public:
   vtkGetMacro(ReplaceActors, int);
   vtkBooleanMacro(ReplaceActors, int);
 
-  virtual void ComputeVisiblePropBounds(vtkRenderer *ren, float bounds[6]);
+  virtual void ComputeVisiblePropBounds(vtkRenderer *ren, double bounds[6]);
 
   // Description:
   // Returns true if remote display is on.  If remote display is on, then
